[ BE ] (perf) Item 클래스 수정
[ BE ] (perf) Item 클래스 수정

int와 double 타입 대신 Integer 타입으로 null를 받아도 에러가 나지 않도록 수정하였습니다.

+++ 4 WEEK/item-service/src/main/java/hello/itemservice/item/Item.java
@@ -8,10 +8,10 @@ import lombok.Setter;
 public class Item {
     private Long id;
     private String name;
-    private double price;
-    private int quantity;
+    private Integer price;
+    private Integer quantity;
 
-    public Item(String name, double price, int quantity){
+    public Item(String name, Integer price, Integer quantity){
         this.name = name;
         this.price = price;
         this.quantity = quantity;
